rename: Home 파일 스타일 분리
